requirements-slim.txt
# Requirements for LiveCold (Pathway + Demo)
flask
flask-socketio
paho-mqtt
litellm
python-dotenv
requests
# Pathway framework (core of the hackathon)
pathway
# Local embeddings (no API calls)
sentence-transformers
tiktoken
